Add CSV cache to currency lambda
